modularized the app by putting the server functions in their own file in the lib folder.
