+++ lib/deleteMemo.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const client = require('./Client');
+
+function deleteMemo(request, response) {
+  let id = request.params.id;
+  const SQL = 'DELETE FROM memos WHERE id=$1;';
+  const VALUES = [id];
+
+  client.query(SQL, VALUES)
+    .then(() => response.redirect('/'))
+    .catch(err => console.log('ya done goofed: ', err));
+}
+
+module.exports = deleteMemo;
